Deleted all tile files, and added three20 to project.
The tile files aren't needed anymore, as we'll be hosting these on a cloudfront tile server. I removed these and added three20 as an git submodule, so we can use it as an external library. This gives us the ability to call tiles off of a tileserver and cache them on the disk, where they'll be available (I believe!) for offline use, after the user quits. Good stuff!